refactor(gateway): refactor healthcheck to use HTTP endpoint
Removes the separate healthcheck binary and exposes a
/api/gateway/health endpoint instead. This shrinks the Docker image and
simplifies the compose healthcheck by using curl. Rename and prepare the
old healthcheck main.go into probe.go for futur use.
