+++ services/gateway/.env.example
@@ -47,4 +47,4 @@ IP_BUCKET=
 #    │     ───────── Example: true ──────────      │
 #    ╰─────────────────────────────────────────────╯
 
-DEV=
+GATEWAY_DEV=
