+++ services/gateway/Dockerfile
@@ -19,8 +19,7 @@ COPY . .
 
 # Disable Go functionality using external C libraries
 # Then tell the compiler to target linux env even if on MacOS or Windows
-RUN CGO_ENABLED=0 GOOS=linux go build -o gateway_binary ./cmd/gateway/ && \
-    CGO_ENABLED=0 GOOS=linux go build -o healthcheck_binary ./cmd/healthcheck/
+RUN CGO_ENABLED=0 GOOS=linux go build -o gateway_binary ./cmd/gateway/
 
 #    ╭────────────────────────────╮
 #    │         Run stage          │
@@ -29,7 +28,6 @@ RUN CGO_ENABLED=0 GOOS=linux go build -o gateway_binary ./cmd/gateway/ && \
 FROM alpine:3.23.3
 RUN apk add --no-cache curl
 COPY --from=build /app/gateway_binary ./gateway_binary
-COPY --from=build /app/healthcheck_binary ./healthcheck_binary
 
 EXPOSE 8080
